convert to staticLib and premake cleanup

+++ bEngine/src/bEngine/ImGui/ImGuiLayer.cpp
@@ -84,7 +84,8 @@ namespace bEngine
         Application& app = Application::Get();
         auto window = static_cast<GLFWwindow*>(app.GetWindow().GetNativeWindow());
 
-        io.DisplaySize = ImVec2(app.GetWindow().GetWidth(), app.GetWindow().GetHeight());
+        io.DisplaySize = ImVec2(static_cast<float>(app.GetWindow().GetWidth()),
+            static_cast<float>(app.GetWindow().GetHeight()));
         
         // Rendering
         ImGui::Render();
